- The fields at JavaBeanUtilTest (DummyBean) must be public (otherwise the java.lang.Class.getFields() won't return the expected fields). - Well, not only the DummyBean's fields must be public, but the whole class has to be public. - Removed the "prefixed" tests. Man, what was I thinking about when doing that?
- The JavaBeanUtil now knows about default fields (those that belongs to the Object class). I implemented
both a method (that discover such fields at runtime) and a static field. We, allover the JavaBeanUtil
class, use only the field: using the method would be much slower. However, there is a test at
JavaBeanUtilTest, the "testIgnoredProperties()", that checks if the field and methods have the same
value. I think it's a good trade off.
- Added a constant to the field string prefix. We need that in the mappingField's methods.
- The mapping methods now ignores Object's default properties.
- Now we correctly guess the property's name from the JavaBeans' methods.
- Removed the 'prefixed' crap.

- Implemented some useful methods at net.sf.ideais.util.ArrayUtil. I'm to used with Python's 'in'
statement, it's a shame Java haven't one. So, I implemented the ArrayUtil.has(), that does
pretty much the same thing.
- Created a method to compare two arrays, ignoring the order of the elements.
- Created a method to duplicate (a shallow dup) an array.